[FIX] web: fix colour for .o_form_uri
Before this revision, multiline links in form views had a different
colour depending on whether we were editing a record, or merely
viewing it.

As an example, the 'customer' field (containing both the name and the
address) in the form view of 'invoice' is affected. When editing the
invoice, all lines have @link-color as colour. However, when viewing it,
only the first line does, while the address is in @odoo-main-text-color.