[FIX] pad: force writing url, even if it did not change
Before this commit, the pad widget did not force the view to save the
current value, since it did not change (pad values are just the url for
the pad).  However, this is a problem for the server, since it uses the
fact that the web client forces a save to update the description field
in the model (get the data from the pad server and add it to the
database).

This was not a big deal when looking at the pad (the pad widget fetches
its data each time), but it is a problem that the database is not
synchronized with the current data.