[REF] web, *: improve behavior of jQuery extension 'prependEvent'
* web_editor

Odoo extends jQuery to add some utility functions. The 'prependEvent'
one add the possibility to add an event handler on an element (like
the jQuery built-in 'on' function) but to be executed before the
already binded ones.

This function was however incomplete and did not allow to use jQuery
namespaced events.