[FIX] stock: add move when picking is available
This completes rev cc3259d8fc3f0f0a0abb21fde970ec27cbf4038c.

The state field was correctly updated by the default_get, however it was
not set to the server because it was set as readonly on the model. For
this specific usecase, we force the readonly to False in the view.

The same issue was happening with the product_uom_id, which is set to
readonly in the view. We mark the field as editable if the move is
additional, and as it does not reserve anything it shouldn't trigger any
other issues.