[FIX] base/module: Use a correct env to load next action on install
Bug introduced in https://github.com/odoo/odoo/commit/c413b49eb48ec6bb91aecf4f01dee7740f6ad3e9

Purpose
=======

If a module installation is supposed to load an action (example: a wizard like in l10n_eu_service) and if the related action is defined in this module (example: l10n_eu_service.wizard), currently we're using the environment defined on self, which is not updated after the module install, and the key is not defined in the registry.

Specification
=============

Instead of using the environment on self, let's use the new environment created after the cursor commit.